+++ instructor/patch.py
@@ -6,7 +6,7 @@ from collections.abc import Iterable
 from functools import wraps
 from tenacity import Retrying, AsyncRetrying, stop_after_attempt, RetryError
 from json import JSONDecodeError
-import xml.etree.ElementTree as ET
+import importlib
 from typing import (
     Callable,
     Optional,
@@ -27,7 +27,6 @@ from openai.types.chat import (
     ChatCompletionMessageParam,
 )
 from openai.types.completion_usage import CompletionUsage
-from anthropic.types.message import Message
 from pydantic import BaseModel, ValidationError
 
 from instructor.dsl.iterable import IterableModel, IterableBase
@@ -36,6 +35,7 @@ from instructor.dsl.partial import PartialBase
 from instructor.dsl.simple_type import ModelAdapter, AdapterBase, is_simple_type
 
 from .function_calls import Mode, OpenAISchema, openai_schema
+from .anthropic_utils import json_to_xml, extract_xml, xml_to_model
 
 logger = logging.getLogger("instructor")
 T = TypeVar("T")
@@ -176,7 +176,7 @@ def handle_response_model(
             else:
                 new_kwargs["messages"][0]["content"] += f"\n\n{message}"
         elif mode == Mode.ANTHROPIC_TOOLS:            
-            tool_descriptions = model_to_xml(response_model)
+            tool_descriptions = json_to_xml(response_model.schema_json())
             system_prompt = f"""In this environment you have access to a set of tools you can use to answer the user's question.
 
                                 You may call them like this:
@@ -430,7 +430,7 @@ def retry_sync(
                 try:
                     response = func(*args, **kwargs)
                     stream = kwargs.get("stream", False)
-                    if isinstance(response, Message): # todo: implement more advanced response handling
+                    if isinstance(response, getattr(importlib.import_module("anthropic.types.message"), "Message")): # todo: implement more advanced response handling
                         return xml_to_model(response_model, extract_xml(response.content[0].text))
                     if (
                         isinstance(response, ChatCompletion)
@@ -493,48 +493,6 @@ def is_async(func: Callable) -> bool:
         func = func.__wrapped__
         is_coroutine = is_coroutine or inspect.iscoroutinefunction(func)
     return is_coroutine
-
-# todo: make function better (edge cases, robustness, etc.)
-def model_to_xml(model: BaseModel) -> str:
-    """Takes a Pydantic model and returns its details in XML format."""
-    root = ET.Element("tool_description")
-    tool_name = ET.SubElement(root, "tool_name")
-    tool_name.text = model.__name__
-    description = ET.SubElement(root, "description")
-    description.text = ("This is the function that must be used to construct the response.")
-    parameters = ET.SubElement(root, "parameters")
-    
-    for field_name, field_type in model.__annotations__.items():
-        parameter = ET.SubElement(parameters, "parameter")
-        
-        name = ET.SubElement(parameter, "name")
-        name.text = field_name
-        
-        type_element = ET.SubElement(parameter, "type")
-        type_element.text = field_type.__name__
-        
-        param_description = ET.SubElement(parameter, "description")
-        param_description.text = f"The {field_name} of the {model.__name__} model"
-    
-    return ET.tostring(root, encoding="unicode")
-
-# todo: make function better (edge cases, robustness, etc.)
-# todo: super primitive parsing, make better
-def extract_xml(content: str) -> str:
-    start_index = content.find('<')
-    end_index = content.rfind('>') + 1
-    return content[start_index:end_index]
-
-# todo: make function better (edge cases, robustness, etc.)
-def xml_to_model(model, xml_string):
-    root = ET.fromstring(xml_string)
-    parameters = {}
-    for param in root.find('.//parameters'):
-        # todo: this assumes all values are strings, fix
-        field_type = model.__annotations__.get(param.tag)
-        if field_type:
-            parameters[param.tag] = field_type(param.text)
-    return model(**parameters)
 
 OVERRIDE_DOCS = """
 Creates a new chat completion for the provided messages and parameters.
